// code/frontend/lib/types.ts
export interface UserSummary {
  _id: string;
  displayName: string;
  avatar?: string | null;
}
// ...
export interface OrgSummary {
  _id: string;
  name: string;
  slug: string;
  avatar?: string | null;
}
// ...
export interface Paper {
  _id: string;
  title: string;
  authors: string[];
  abstract: string | null;
  keywords: string[];
  doi: string | null;
  isbn?: string | null;
  publicationDate?: string | null;
  year: number | null;
  journal: string | null;
  fileUrl: string | null;
  fileSize: number | null;
  uploadedBy: UserSummary;
  organizationId?: OrgSummary | null;
  sourcePostId?: string | null;
  isPublished: boolean;
  downloadCount: number;
  topics?: string[];
  createdAt: string;
  updatedAt: string;
}